fix warnings

+++ include/ads/ads-vocab.hpp
@@ -42,7 +42,7 @@ struct frame_idx     { int64_t value = 0; };
 [[nodiscard]] constexpr inline auto operator<=>(channel_idx lhs, channel_count rhs)               { return lhs.value <=> rhs.value; }
 [[nodiscard]] constexpr inline auto operator<=>(channel_idx lhs, channel_idx rhs)                 { return lhs.value <=> rhs.value; }
 [[nodiscard]] constexpr inline auto operator<=>(concepts::arithmetic auto lhs, frame_count rhs)   { return lhs <=> rhs.value; }
-[[nodiscard]] constexpr inline auto operator<=>(concepts::arithmetic auto lhs, frame_idx rhs)     { return lhs <=> rhs.value; }
+[[nodiscard]] constexpr inline auto operator<=>(concepts::arithmetic auto lhs, frame_idx rhs)     { return static_cast<int64_t>(lhs) <=> rhs.value; }
 [[nodiscard]] constexpr inline auto operator<=>(frame_count lhs, concepts::arithmetic auto rhs)   { return lhs.value <=> rhs; }
 [[nodiscard]] constexpr inline auto operator<=>(frame_count lhs, frame_count rhs)                 { return lhs.value <=> rhs.value; }
 [[nodiscard]] constexpr inline auto operator<=>(frame_count lhs, frame_idx rhs)                   { return static_cast<int64_t>(lhs.value) <=> rhs.value; }
